requirement.txt
gradio>=4.0.0
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
